Move pool impls to its own module
